Support for context switch and moved stack management APIS to score
These functions should be called during context switch, initialization and restore
