+++ cpukit/headers.am
@@ -16,7 +16,6 @@ include_HEADERS += include/mqueue.h
 include_HEADERS += include/pci.h
 include_HEADERS += include/poll.h
 include_HEADERS += include/rtems.h
-include_HEADERS += include/stack_management.h
 include_HEADERS += include/sha256.h
 include_HEADERS += include/sha512.h
 include_HEADERS += include/t.h
@@ -407,6 +406,7 @@ include_rtems_score_HEADERS += include/rtems/score/smplockstats.h
 include_rtems_score_HEADERS += include/rtems/score/smplockticket.h
 include_rtems_score_HEADERS += include/rtems/score/stack.h
 include_rtems_score_HEADERS += include/rtems/score/stackimpl.h
+include_rtems_score_HEADERS += include/rtems/score/stackmanagement.h
 include_rtems_score_HEADERS += include/rtems/score/states.h
 include_rtems_score_HEADERS += include/rtems/score/statesimpl.h
 include_rtems_score_HEADERS += include/rtems/score/status.h
